Don't need `my_model = None` to clear old model
Since we're going to update `my_model` a few lines later (in both cases) and we
are not going to use the current value of `my_model` before that line anyway,
there's no need to first set it to `None` before setting it to the output of
`build_model()` on the next line or a few lines later.

This simplifies the code and improves readability.